Move mock_executor_run_command to test utilities
